requirements.txt
annotated-types==0.6.0
anyio==3.7.0
bcrypt==4.0.1
certifi==2024.2.2
cffi==1.16.0
click==8.1.3
cryptography==42.0.5
ecdsa==0.18.0
fastapi==0.98.0
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.5
httptools==0.5.0
httpx==0.27.0
idna==3.4
passlib==1.7.4
pyasn1==0.5.0
pycparser==2.22
pydantic==1.10.9
pydantic_core==2.18.2
PyMySQL==1.1.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.17
starlette==0.27.0
typing_extensions==4.6.3
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
